feat: added Spike inside invisible box
